rpi5/requirements.txt
requests
adafruit-blinka
pillow
numpy
click
Adafruit-Blinka-Raspberry-Pi5-Piomatter
adafruit-circuitpython-matrixportal
